added User Authetnication features

+++ LittleLemon/resturant/models.py
@@ -9,4 +9,4 @@ class Booking (models.Model):
 class MenuItems (models.Model):
     Title = models.CharField(max_length=255)
     Price = models.DecimalField(max_digits=10, decimal_places=2)
-    Inventory = models.IntegerField()
+    Inventory = models.SmallIntegerField()
